Fixes declaration order issue #86
I moved things around. Also pulled the plugin_loaded stuff out
of its function since it isn't used anywhere. I'll admit that
I don't know if this function is required as part of the sublime
API or anything, but I suspect its not, so I figured, while I
was at it, I might as well save a few lines.